1.reuse runderivation code 2.revert hostio functions to start with lowercase 3.fix doc error
