+++ op-program/client/full_run_conf.go
@@ -1,0 +1,4 @@
+//go:build !smoke_test
+package client
+
+var maximumSteps = -1
