+++ op-program/client/smoke_test_conf.go
@@ -1,0 +1,6 @@
+//go:build smoke_test
+
+package client
+
+// in smoke test we only run 10 steps
+var maximumSteps = 10
